Store github test -repo and -owner in config file
This makes it easier for new contributors to run the tests. I tried to
run the tests according to the contribution guidelines in `README.md`.
Unfortunately the tests failed. From the error messages I could tell
that the repository and github account used for the tests are hard-coded
which is bad for more than one contributor.

By the way, this is my very first commit to a golang project. Any
feedback welcome! :heart: